requirements.txt
streamlit
langchain
langchain-groq
langchain-community
chromadb
pypdf
sentence-transformers
python-dotenv
pymupdf